+++ src/connector/src/parser/dynamodb/cdc_json_parser.rs
@@ -12,16 +12,84 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use assert_matches::assert_matches;
+use risingwave_common::types::{DataType, DatumCow, ScalarRefImpl, ToDatumRef};
+
 use crate::error::ConnectorResult;
 use crate::only_parse_payload;
-use crate::parser::dynamodb::{build_dynamodb_json_accessor_builder, DynamodbChangeEvent};
+use crate::parser::dynamodb::{build_dynamodb_json_accessor_builder, map_rw_type_to_dynamodb_type};
 use crate::parser::unified::util::apply_row_operation_on_stream_chunk_writer;
+use crate::parser::unified::{ChangeEvent, ChangeEventOperation};
 use crate::parser::{
-    AccessBuilderImpl, ByteStreamSourceParser, ParserFormat, SourceStreamChunkRowWriter,
-    SpecificParserConfig,
+    Access, AccessBuilderImpl, AccessError, ByteStreamSourceParser, ParserFormat,
+    SourceStreamChunkRowWriter, SpecificParserConfig,
 };
 use crate::source::{SourceColumnDesc, SourceContext, SourceContextRef};
 
+// See https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_streams_Record.html
+// An example of a DynamoDB change event from Kinesis:
+// {
+//     "awsRegion":"us-east-1",
+//     "eventID":"26b1b865-1ee7-47f9-816e-d51cdeef9dc5",
+//     "eventName":"MODIFY",
+//     "userIdentity":null,
+//     "recordFormat":"application/json",
+//     "tableName":"wkx-test-orders",
+//     "dynamodb":{
+//         "ApproximateCreationDateTime":1720046486486729,
+//         "Keys":{
+//             "customer_name":{
+//                 "S":"Bob Lee"
+//             },
+//             "order_id":{
+//                 "N":"3"
+//             }
+//         },
+//         "NewImage":{
+//             "order_status":{
+//                 "N":"3"
+//             },
+//             "order_date":{
+//                 "N":"1720046486"
+//             },
+//             "order_id":{
+//                 "N":"3"
+//             },
+//             "price":{
+//                 "N":"63.06"
+//             },
+//             "product_id":{
+//                 "N":"2060"
+//             },
+//             "customer_name":{
+//                 "S":"Bob Lee"
+//             }
+//         },
+//         "OldImage":{
+//             "order_status":{
+//                 "N":"3"
+//             },
+//             "order_date":{
+//                 "N":"1720037677"
+//             },
+//             "order_id":{
+//                 "N":"3"
+//             },
+//             "price":{
+//                 "N":"63.06"
+//             },
+//             "product_id":{
+//                 "N":"2060"
+//             },
+//             "customer_name":{
+//                 "S":"Bob Lee"
+//             }
+//         },
+//         "SizeBytes":192,
+//         "ApproximateCreationDateTimePrecision":"MICROSECOND"
+//     },
+//     "eventSource":"aws:dynamodb"
+// }
 #[derive(Debug)]
 pub struct DynamodbCdcJsonParser {
     payload_builder: AccessBuilderImpl,
@@ -81,5 +149,81 @@ impl ByteStreamSourceParser for DynamodbCdcJsonParser {
         writer: SourceStreamChunkRowWriter<'a>,
     ) -> ConnectorResult<()> {
         only_parse_payload!(self, payload, writer)
+    }
+}
+
+struct DynamodbChangeEvent<A> {
+    value_accessor: A,
+    single_blob_column: String,
+}
+
+const OLD_IMAGE: &str = "OldImage";
+const NEW_IMAGE: &str = "NewImage";
+const DYNAMODB: &str = "dynamodb";
+
+const DYNAMODB_CREATE_OP: &str = "INSERT";
+const DYNAMODB_MODIFY_OP: &str = "MODIFY";
+const DYNAMODB_REMOVE_OP: &str = "REMOVE";
+
+const OP: &str = "eventName";
+
+impl<A> DynamodbChangeEvent<A>
+where
+    A: Access,
+{
+    pub fn new(value_accessor: A, single_blob_column: String) -> Self {
+        Self {
+            value_accessor,
+            single_blob_column,
+        }
+    }
+}
+
+impl<A> ChangeEvent for DynamodbChangeEvent<A>
+where
+    A: Access,
+{
+    fn access_field(&self, desc: &SourceColumnDesc) -> crate::parser::AccessResult<DatumCow<'_>> {
+        if desc.name == self.single_blob_column {
+            assert_matches!(desc.data_type, DataType::Jsonb);
+            match self.op()? {
+                ChangeEventOperation::Delete => self
+                    .value_accessor
+                    .access(&[DYNAMODB, OLD_IMAGE], &desc.data_type),
+                ChangeEventOperation::Upsert => self
+                    .value_accessor
+                    .access(&[DYNAMODB, NEW_IMAGE], &desc.data_type),
+            }
+        } else {
+            let dynamodb_type = map_rw_type_to_dynamodb_type(&desc.data_type)?;
+            match self.op()? {
+                ChangeEventOperation::Delete => self.value_accessor.access(
+                    &[DYNAMODB, OLD_IMAGE, &desc.name, dynamodb_type.as_str()],
+                    &desc.data_type,
+                ),
+                ChangeEventOperation::Upsert => self.value_accessor.access(
+                    &[DYNAMODB, NEW_IMAGE, &desc.name, dynamodb_type.as_str()],
+                    &desc.data_type,
+                ),
+            }
+        }
+    }
+
+    fn op(&self) -> Result<ChangeEventOperation, AccessError> {
+        if let Some(ScalarRefImpl::Utf8(op)) = self
+            .value_accessor
+            .access(&[OP], &DataType::Varchar)?
+            .to_datum_ref()
+        {
+            match op {
+                DYNAMODB_CREATE_OP | DYNAMODB_MODIFY_OP => return Ok(ChangeEventOperation::Upsert),
+                DYNAMODB_REMOVE_OP => return Ok(ChangeEventOperation::Delete),
+                _ => panic!("Unknown dynamodb event operation: {}", op),
+            }
+        }
+        Err(super::AccessError::Undefined {
+            name: "op".into(),
+            path: String::from(OP),
+        })
     }
 }
